+++ src/Infrastructure/Handler/IRepositoryCommand.cs
@@ -1,9 +1,12 @@
+using System.Threading;
 using System.Threading.Tasks;
+using ViajaNet.JobApplication.Application.Core;
 
 namespace ViajaNet.JobApplication.Infrastructure
 {
     public interface IRepositoryCommand
     {
-        Task<string> CreateAsync();
+        Task<long> CreateAsync(AnalyticsEntity entity);
+        Task<long> CreateAsync(AnalyticsEntity entity, CancellationToken cancellationToken);
     }
 }