+++ src/Infrastructure/Handler/IRepositoryQuery.cs
@@ -1,3 +1,5 @@
+using System.Collections.Generic;
+using System.Threading;
 using System.Threading.Tasks;
 using ViajaNet.JobApplication.Application.Core;
 
@@ -5,6 +7,9 @@ namespace ViajaNet.JobApplication.Infrastructure
 {
     public interface IRepositoryQuery
     {
-        Task<AnalyticsDto> QueryAsync();
+        Task<AnalyticsEntity> QueryByIdAsync(long id);
+        Task<AnalyticsEntity> QueryByIdAsync(long id, CancellationToken cancellationToken);
+        Task<IEnumerable<AnalyticsEntity>> QueryByParametersAsync(string ip, string pageName);
+        Task<IEnumerable<AnalyticsEntity>> QueryByParametersAsync(string ip, string pageName, CancellationToken cancellationToken);
     }
 }